Moved copy, paste and delete to its own functions
Also added storing the current modifier key for later
