+++ client/src/game/events/keyboard.ts
@@ -1,44 +1,44 @@
 import Tools from "@/game/ui/tools/tools.vue";
 
-import { getRef, uuidv4 } from "@/core/utils";
+import { getRef } from "@/core/utils";
 import { socket } from "@/game/api/socket";
 import { sendClientOptions } from "@/game/api/utils";
-import { ServerAura } from '@/game/comm/types/shapes';
-import { EventBus } from "@/game/event-bus";
 import { Vector } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
-import { createShapeFromDict } from "@/game/shapes/utils";
 import { gameStore } from "@/game/store";
 import { calculateDelta } from "@/game/ui/tools/utils";
+import { copyShapes, deleteShapes, pasteShapes } from '../shapes/utils';
+
+export const keyboardState = {
+    shiftPressed: false,
+    ctrlPressed: false,
+    altPressed: false,
+}
+
+function updateKeyboardState(event: KeyboardEvent) {
+    keyboardState.shiftPressed = event.getModifierState("Shift")
+    keyboardState.ctrlPressed = event.getModifierState("Control")
+    keyboardState.altPressed = event.getModifierState("Alt")
+}
 
 export function onKeyUp(event: KeyboardEvent) {
+    updateKeyboardState(event);
     if (event.target instanceof HTMLInputElement || event.target instanceof HTMLTextAreaElement) {
     } else {
         if (event.key === "Delete" || event.key === "Del" || event.key === "Backspace") {
-            if (layerManager.getLayer === undefined) {
-                console.log("No active layer selected for delete operation");
-                return;
-            }
-            const l = layerManager.getLayer()!;
-            for (let i = l.selection.length - 1; i >= 0; i--) {
-                const sel = l.selection[i];
-                if ((<any>getRef<Tools>("tools").$refs.selectTool).selectionHelper.uuid === sel.uuid) {
-                    l.selection.splice(i, 1);
-                    continue;
-                }
-                l.removeShape(sel, true, false);
-                EventBus.$emit("SelectionInfo.Shape.Set", null);
-                EventBus.$emit("Initiative.Remove", sel.uuid);
-            }
+            deleteShapes()
         }
     }
 }
 
 export function onKeyDown(event: KeyboardEvent) {
+    updateKeyboardState(event);
     if (event.target instanceof HTMLInputElement || event.target instanceof HTMLTextAreaElement) {
-        if (event.keyCode === 65 && event.ctrlKey) event.target!.select();
+        // Ctrl-a with a HTMLInputElement or a HTMLTextAreaElement selected - select all the text
+        if (event.key === "a" && event.ctrlKey) event.target!.select();
     } else {
         if (event.keyCode >= 37 && event.keyCode <= 40) {
+            // Arrow keys - move the selection or the camera
             // todo: this should already be rounded
             const gridSize = Math.round(gameStore.gridSize);
             let offsetX = gridSize * (event.keyCode % 2);
@@ -71,61 +71,24 @@ export function onKeyDown(event: KeyboardEvent) {
                 layerManager.invalidate();
                 sendClientOptions();
             }
-        } else if (event.keyCode === 68) {
+        } else if (event.key === "d") {
+            // d - Deselect all
             const layer = layerManager.getLayer();
             if (layer) {
                 layer.clearSelection();
                 layer.invalidate(true);
             }
         } else if (event.key === "u" && event.ctrlKey) {
+            // Ctrl-u - disable and reenable the Interface
             event.preventDefault();
             event.stopPropagation();
             gameStore.toggleUI();
         } else if (event.key === "c" && event.ctrlKey) {
-            const layer = layerManager.getLayer();
-            if (!layer) return;
-            if (!layer.selection) return;
-            const clipboard = [];
-            for (const shape of layer.selection) {
-                if ((<any>getRef<Tools>("tools").$refs.selectTool).selectionHelper.uuid === shape.uuid) continue;
-                clipboard.push(shape.asDict());
-            }
-            gameStore.setClipboard(clipboard);
+            // Ctrl-c - Copy
+            copyShapes()
         } else if (event.key === "v" && event.ctrlKey) {
-            const layer = layerManager.getLayer();
-            if (!layer) return;
-            if (!gameStore.clipboard) return;
-            layer.selection = [];
-            for (const clip of gameStore.clipboard) {
-                clip.x += 10;
-                clip.y += 10;
-                clip.uuid = uuidv4();
-                const oldTrackers = clip.trackers;
-                clip.trackers = [];
-                for (const tracker of oldTrackers) {
-                    const newTracker: Tracker = {
-                        ...tracker,
-                        uuid: uuidv4()
-                    }
-                    clip.trackers.push(newTracker);
-                }
-                const oldAuras = clip.auras;
-                clip.auras = [];
-                for (const aura of oldAuras) {
-                    const newAura: ServerAura = {
-                        ...aura,
-                        uuid: uuidv4()
-                    }
-                    clip.auras.push(newAura);
-                }
-                const shape = createShapeFromDict(clip);
-                if (shape === undefined) continue;
-                layer.addShape(shape, true);
-                layer.selection.push(shape);
-            }
-            if (layer.selection.length === 1) EventBus.$emit("SelectionInfo.Shape.Set", layer.selection[0]);
-            else EventBus.$emit("SelectionInfo.Shape.Set", null);
-            layer.invalidate(false);
+            // Ctrl-v - Paste
+            pasteShapes()
         }
     }
 }
